requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.7.2
    # via flask
blinker==1.6.2
    # via flask
click==8.1.7
    # via flask
flask[async]==2.3.3
    # via
    #   -r requirements.in
    #   flask-cors
flask-cors==4.0.0
    # via -r requirements.in
geode-background==7.3.7
    # via
    #   geode-explicit
    #   geode-implicit
    #   geode-simplex
geode-common==26.2.0
    # via
    #   -r requirements.in
    #   geode-background
    #   geode-explicit
    #   geode-implicit
    #   geode-simplex
    #   geode-viewables
geode-conversion==5.0.14
    # via
    #   geode-explicit
    #   geode-implicit
geode-explicit==4.2.1
    # via
    #   -r requirements.in
    #   geode-implicit
geode-implicit==2.2.1
    # via -r requirements.in
geode-numerics==3.0.11
    # via
    #   -r requirements.in
    #   geode-implicit
    #   geode-simplex
geode-simplex==6.1.7
    # via -r requirements.in
geode-viewables==2.0.5
    # via opengeodeweb-back
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
markupsafe==2.1.3
    # via
    #   jinja2
    #   werkzeug
opengeode-core==14.6.3
    # via
    #   -r requirements.in
    #   geode-background
    #   geode-common
    #   geode-conversion
    #   geode-explicit
    #   geode-implicit
    #   geode-numerics
    #   geode-simplex
    #   geode-viewables
    #   opengeode-geosciences
    #   opengeode-geosciencesio
    #   opengeode-inspector
    #   opengeode-io
    #   opengeodeweb-back
opengeode-geosciences==7.1.8
    # via
    #   -r requirements.in
    #   geode-explicit
    #   geode-implicit
    #   geode-viewables
    #   opengeode-geosciencesio
    #   opengeode-inspector
    #   opengeodeweb-back
opengeode-geosciencesio==4.3.2
    # via
    #   -r requirements.in
    #   geode-explicit
    #   geode-implicit
    #   opengeode-inspector
    #   opengeodeweb-back
opengeode-inspector==3.0.11
    # via
    #   -r requirements.in
    #   geode-explicit
    #   geode-implicit
    #   opengeodeweb-back
opengeode-io==6.1.1
    # via
    #   -r requirements.in
    #   geode-explicit
    #   geode-implicit
    #   geode-viewables
    #   opengeode-inspector
    #   opengeodeweb-back
opengeodeweb-back==0.0.15
    # via -r requirements.in
python-dotenv==1.0.0
    # via -r requirements.in
typing-extensions==4.7.1
    # via asgiref
werkzeug==2.3.7
    # via
    #   -r requirements.in
    #   flask
